@@ -28,8 +28,7 @@ class PrepareImages {
         this.attachmentRepository = attachmentRepository;
 
         for (Attachment att : message.getAttachments()) {
-            if (att instanceof FileAttachment) {
-                FileAttachment fatt = (FileAttachment) att;
+            if (att instanceof FileAttachment fatt) {
                 if (fatt.getContentId() != null) {
                     attachmentById.put(fatt.getContentId(), fatt);
                 }
