@@ -1,11 +1,4 @@
 package net.sourceforge.MSGViewer;
 
-class Content {
-    public final String type;
-    public final String text;
-
-    Content(String type, String text) {
-        this.type = type;
-        this.text = text;
-    }
+record Content(String type, String text) {
 }
