@@ -42,19 +42,23 @@ public class ModuleLauncher extends BaseModuleLauncher {
             CLIHelpMSGViewer help = new CLIHelpMSGViewer(this);
             help.printVersion();
 
-            System.out.println("Copyright (C) 2015  Martin Oberzalek <msgviewer@hoffer.cx>\n" +
-                    "\n" +
-                    "This program is free software; you can redistribute it and/or modify\n" +
-                    "it under the terms of the GNU General Public License as published by\n" +
-                    "the Free Software Foundation; either version 3 of the License, or\n" +
-                    "(at your option) any later version.   \n\n" +
-                    "This program is distributed in the hope that it will be useful,\n" +
-                    "but WITHOUT ANY WARRANTY; without even the implied warranty of\n" +
-                    "MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the\n" +
-                    "GNU General Public License for more details.\n\n" +
-                    "You should have received a copy of the GNU General Public License\n" +
-                    "along with this program; if not, write to the Free Software Foundation,\n" +
-                    "Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301  USA\n");
+            System.out.println("""
+                    Copyright (C) 2015  Martin Oberzalek <msgviewer@hoffer.cx>
+
+                    This program is free software; you can redistribute it and/or modify
+                    it under the terms of the GNU General Public License as published by
+                    the Free Software Foundation; either version 3 of the License, or
+                    (at your option) any later version.  \s
+
+                    This program is distributed in the hope that it will be useful,
+                    but WITHOUT ANY WARRANTY; without even the implied warranty of
+                    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+                    GNU General Public License for more details.
+
+                    You should have received a copy of the GNU General Public License
+                    along with this program; if not, write to the Free Software Foundation,
+                    Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301  USA
+                    """);
             return;
         }
 
