@@ -27,7 +27,7 @@ public abstract class RecipientHeader {
     protected Stream<RecipientEntry> splitAttendees(String text) {
         return Arrays.stream(text.split(","))
                 .map(RecipientHeader::mailAddressFrom)
-                .filter(addr -> addr.getEmail().contains("@"))
+                .filter(addr -> addr.email().contains("@"))
                 .map(this::toRecipientEntry);
     }
 
@@ -45,8 +45,8 @@ public abstract class RecipientHeader {
 
     private RecipientEntry toRecipientEntry(MailAddress email) {
         RecipientEntry recipientEntry = new RecipientEntry();
-        recipientEntry.setEmail(email.getEmail());
-        recipientEntry.setName(email.getDisplayName());
+        recipientEntry.setEmail(email.email());
+        recipientEntry.setName(email.displayName());
         recipientEntry.setType(type);
         return recipientEntry;
     }
